+++ internal/docxwrappers/fumiama_docx_wrapper.go
@@ -1,4 +1,4 @@
-package docx_wrappers
+package docxwrappers
 
 import (
 	"archive/zip"
